Move crypto3 derivation into a separate file, and clean it up
This diff introduces the following changes:

- crypto3 derivation moved into a separate file, as per usual Nix guidelines
- enableDebug parameter added to crypto3 to switch debug modes on and off
- enableDebug will build boost with debug symbols
- flake-utils used to enumerate systems (as we do in nix-3rdparty)
- devShells dropped from the toplevel. please use shellHoook in crypto3.nix instead

This is all a set of standard practices to make your life easier and
code cleaner.

What still remains is to remove the hardcoding of test sets in the
flake.nix. This usually is not recommended to program a lot of
logic. The best place to do this would be in a shell script.